amount is int

+++ src/main/java/at/hannibal2/skyhanni/events/yearofthepig/ShinyOrbLootedEvent.kt
@@ -29,7 +29,7 @@ class ShinyOrbLootedEvent(
 
         data class SkillXp(
             val skill: SkillType,
-            val amount: Long,
+            val amount: Int,
         ) : Reward
     }
 }
